Use FloatField for the best performance

+++ server/models.py
@@ -79,7 +79,7 @@ class Sensor(models.Model):
 
 class SensorValue(models.Model):
     sensor = models.ForeignKey('Sensor')
-    value = models.DecimalField(max_digits=6, decimal_places=2)
+    value = models.FloatField()
     timestamp = models.DateTimeField(auto_now=False, db_index=True)
 
     def __unicode__(self):
